refactor: move RenderAlreadyLoadedGeometryProvider
