+++ viewer/core/src/datamodels/cad/createCadManager.ts
@@ -5,9 +5,9 @@
 import { CadManager } from './CadManager';
 import { CadModelFactory } from './CadModelFactory';
 
-import { CadModelUpdateHandler, createV8SectorCuller, OccludingGeometryProvider } from '@reveal/cad-geometry-loaders';
+import { CadModelUpdateHandler, createV8SectorCuller } from '@reveal/cad-geometry-loaders';
 
-import { CadMaterialManager } from '@reveal/rendering';
+import { CadMaterialManager, EffectRenderManager } from '@reveal/rendering';
 
 import { RevealOptions } from '../../public/types';
 
@@ -18,7 +18,7 @@ export function createCadManager(
   modelDataProvider: ModelDataProvider,
   renderer: THREE.WebGLRenderer,
   materialManager: CadMaterialManager,
-  occludingGeometryProvider: OccludingGeometryProvider,
+  renderManager: EffectRenderManager,
   options: RevealOptions
 ): CadManager {
   const cadModelFactory = new CadModelFactory(materialManager, modelMetadataProvider, modelDataProvider);
@@ -26,7 +26,7 @@ export function createCadManager(
   const sectorCuller =
     internal && internal.sectorCuller
       ? internal.sectorCuller
-      : createV8SectorCuller(renderer, occludingGeometryProvider);
+      : createV8SectorCuller(renderer, renderManager);
   const cadModelUpdateHandler = new CadModelUpdateHandler(sectorCuller, options.continuousModelStreaming);
   return new CadManager(materialManager, cadModelFactory, cadModelUpdateHandler);
 }
