+++ viewer/packages/cad-geometry-loaders/index.ts
@@ -16,5 +16,4 @@ export { ByScreenSizeSectorCuller } from './src/sector/culling/ByScreenSizeSecto
 
 export { createV8SectorCuller } from './src/sector/culling/createV8SectorCuller';
 
-export { OccludingGeometryProvider } from './src/sector/culling/OccludingGeometryProvider';
 export { DetermineSectorsInput } from './src/sector/culling/types';
